requirements.txt
pillow>=9.0.0
transformers>=4.26.0
torch>=1.12.0
timm>=0.6.0 